requirements-cuda.txt
# ValonyLabs Studio v3.0 — CUDA stack (RTX 30/40/50, Colab, Brev)
# Install with: pip install -r requirements-cuda.txt
# Assumes a matching torch+cuda wheel is already installed for your GPU arch.

# ── Core ────────────────────────────────────────────────────────
pydantic>=2.6
fastapi>=0.110
uvicorn[standard]>=0.29
pyyaml>=6.0
rich>=13.7
typer>=0.12

# ── HF / training ───────────────────────────────────────────────
transformers>=4.44
accelerate>=0.30
peft>=0.11
trl>=0.12
datasets>=2.19
huggingface_hub>=0.23
sentencepiece>=0.2
tiktoken>=0.7
safetensors>=0.4
bitsandbytes>=0.43
# Unsloth (Linux-only wheels; Windows users: use WSL2)
unsloth @ git+https://github.com/unslothai/unsloth.git ; sys_platform == 'linux'
xformers>=0.0.27 ; sys_platform == 'linux'
# Flash Attention 2/3 (Ampere+). Comment out on Turing / T4.
flash-attn>=2.6 ; sys_platform == 'linux'

# ── Inference ───────────────────────────────────────────────────
vllm>=0.6.0
# Optional: uncomment to add SGLang
# sglang[all]>=0.3.0

# ── Data Forge ──────────────────────────────────────────────────
pymupdf>=1.24
pdfplumber>=0.11
python-docx>=1.1
python-pptx>=0.6.23
openpyxl>=3.1
pandas>=2.2
chardet>=5.2
beautifulsoup4>=4.12
trafilatura>=1.12
pillow>=10.3
numpy>=1.26

# ── OCR ─────────────────────────────────────────────────────────
rapidocr-onnxruntime>=1.3
paddleocr>=2.7 ; sys_platform == 'linux'
paddlepaddle-gpu>=2.6 ; sys_platform == 'linux'
docling>=1.8 ; sys_platform == 'linux'
pytesseract>=0.3.10

# ── UI ──────────────────────────────────────────────────────────
gradio>=4.44
